Add an documentation to the simple SplitEquippable
This commit adds a documentation to simple SplitEquippable as well as
modifies its deploy script to simply deploy it. User journey script has
been added to replace the showcase of simple SplitEquippable usage.

Custom script for running the simple SplitEquippable user journey has
been added.

Simple SplitEquippable example smart contracts have been moved into
their own folder, so that the documentation of each architecture can be
separate from the main README.
